btc_sniper/requirements.txt
# ═══ Polymarket BTC Sniper v2.3 — Dependencies ═══
# Pin minimum versions for reproducibility

websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
rich>=13.7.0
web3>=6.15.0
py-clob-client>=0.34.5
aioconsole>=0.6.0
pydantic
pytest>=8.0.0
pytest-asyncio>=0.23.0
